Remove custom classes in the favor of factory method in LayerEditor example
